Реализовал методы: - изменение величины испытательного срока; - проверка о прохождении испытательного срока и отправка сообщения.

+++ src/main/java/com/skypro/sharehome/repository/ShareHomeRepository.java
@@ -10,13 +10,13 @@ import org.springframework.stereotype.Repository;
 public interface ShareHomeRepository extends JpaRepository<ShareHome, Long> {
 
     @Query(value = "SELECT description FROM share_home where id = :id", nativeQuery = true)
-    String getDescriptionShareHome(@Param("id") Integer id);
+    String getDescriptionShareHome(@Param("id") Long id);
 
     @Query(value = "SELECT name FROM share_home where id = :id", nativeQuery = true)
-    String getNameShareHome(@Param("id") Integer id);
+    String getNameShareHome(@Param("id") Long id);
 
     @Query(value = "SELECT address FROM share_home WHERE id = :id", nativeQuery = true)
-    String getAddressShareHome(@Param("id") Integer id);
+    String getAddressShareHome(@Param("id") Long id);
 
     @Query(value = "SELECT security FROM share_home WHERE id = :id", nativeQuery = true)
     String getSecurityShareHome(@Param("id") Long id);
